requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
regex==2024.5.15
pandas==2.2.2
python-docx==1.1.2
pymupdf==1.24.7
